added benchmark files, corrected mistakes

+++ pointcloud/run/zand/loose.py
@@ -1,30 +1,42 @@
 # -*- coding: utf-8 -*-
 """
-Created on Sun Mar 13 11:26:53 2016
+Created on Mon Apr 04 14:38:34 2016
 
 @author: Stella Psomadaki
 """
+
 from pointcloud.AbstractBulkLoader import BulkLoader
 from pointcloud.AbstractQuerier import Querier
 import time
 from tabulate import tabulate
 import pointcloud.oracleTools as ora
-#import os
+import os
 
+###########################
+dataset = 'zandmotor'
+integrations = ['lxyt', 'lxyzt']
+scalings = ['1']
+repeatQueries = 6
+###########################
+
+if dataset == 'zandmotor':
+    bench = 3
+elif dataset == 'coastline':
+    bench = 4
+    
+
+path = os.getcwd()
 benchmark = ['mini', 'medium', 'full']  
 hloading = ['approach', 'preparation', 'loading', 'closing', 'size[MB]', 'points']
-
-integrations = ['dxyt']
-scalings = ['10000']
 
 
 for integr in integrations:
     for scaling in scalings:
         loadings = []
         queries = []
-        for i in range(1,4):
-            configuration = 'D:/Dropbox/Thesis/Thesis/Code/ini/zandmotor/{0}_{1}_part{2}.ini'.format(integr, scaling, i)
-            
+        for i in range(1,bench + 1):
+            configuration = path + '/ini/' + dataset + '/' + integr + '_' + scaling + '_part' + str(i) + '.ini'
+
             bulk = BulkLoader(configuration)
             loading = []
             loading.append(benchmark[i - 1])
@@ -47,18 +59,21 @@ for integr in integrations:
             
             loadings.append(loading)
             print tabulate(loadings, hloading, tablefmt="plain")
-            
-#            if i == 1:
-#                os.system('python -m pointcloud.queryTab {0}'.format(configuration)) 
-                
+
             hquery =  ["id", "prep.", 'insert', 'ranges', 'fetching', "decoding", 'storing', "Appr.pts", "Fin.pts", "FinFilt", "time", 'extra%', 'total']
         
             querier = Querier(configuration)
             connection = bulk.getConnection()
             cursor = connection.cursor()
-            print 'querying...'
+            
+            cursor.execute('SELECT table_name FROM all_tables WHERE table_name = :1',[querier.queriesTable.upper(),])
+            length = len(cursor.fetchall())
+            connection.close()
+            if not length:
+                os.system('python -m pointcloud.queryTab {0}'.format(configuration))
+
             for num in querier.ids:
-                for j in range(5):
+                for j in range(repeatQueries):
                     start = time.time()
                     lst = querier.query(num)
                     lst.append(round(time.time() - start, 2))
@@ -67,17 +82,21 @@ for integr in integrations:
                     lst.insert(0, num)
                     queries.append(lst)
                     ora.dropTable(cursor, querier.queryTable + '_' +  str(num))               
-                ora.dropTable(cursor, querier.rangeTable + str(num))
-            
+                if querier.integration == 'deep':
+                    ora.dropTable(cursor, querier.rangeTable + str(num))
+            print
             print tabulate(queries, hquery, tablefmt="plain")
-        
-        print 'writing to file statistics for case: {0}_{1}_{2}'.format(bulk.integration, bulk.parse, bulk.scale)
-        f = open('loading_stats_31_03_load_2.txt', 'a')
-        f.write('Statistics for case: {0} {1} {2}\n'.format(bulk.integration, bulk.parse, bulk.scale))
-        f.write(tabulate(loadings, hloading, tablefmt="plain"))
-        f.write('\n')
-        f.write(tabulate(queries, hquery, tablefmt="plain"))
-        f.write('\n')
-        f.write('\n')
-        f.write('\n')
-        f.close()
+
+    
+    print 'Writing file with results for case: {0} {1} {2}'.format(bulk.integration, bulk.parse, bulk.scale)
+    f = open(path + '/results/{0}/{1}.txt'.format(dataset, integr), 'a')
+    f.write('Statistics for case: {0} {1} {2}\n'.format(bulk.integration, bulk.parse, bulk.scale))
+    f.write('Loading')
+    f.write(tabulate(loadings, hloading, tablefmt="plain"))
+    f.write('\n')
+    f.write('Querying')
+    f.write(tabulate(queries, hquery, tablefmt="plain"))
+    f.write('\n')
+    f.write('\n')
+    f.write('\n')
+    f.close()
